requirements.txt
api==0.0.6
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
dnspython==1.15.0
dnspython3==1.15.0
future==0.16.0
futures==3.1.1
geoip2==2.5.0
geonames==0.1.3
geopy==1.11.0
html5lib==0.999999999
idna==2.6
ipwhois==1.0.0
lxml==4.0.0
maxminddb==1.3.0
maxminddb-geolite2==2017.803
nose==1.3.7
python-builtwith==0.2.2
python-geoip==1.2
python-geoip-geolite2==2015.303
python-whois==0.6.5
pytz==2017.2
requests==2.18.4
seolib==0.1.3
six==1.11.0
unicodecsv==0.14.1
urllib3==1.22
webencodings==0.5.1